feat: update dependabot configuration for pip, devcontainers, and docker (#50)
This pull request primarily updates the `.github/dependabot.yml` file to enhance the configuration of Dependabot, a tool used for dependency management. The changes include expanding the scope of directories for the `pip` and `devcontainers` package ecosystems to all directories, adding a new `docker` package ecosystem, and specifying update types to ignore for the `docker` ecosystem.